* Updates all dependencies * Moved to tsdx project setup * Moved to yarn * Validator.tsx: Fixed typescript errors
