+++ examples/formik/package.json
@@ -3,34 +3,25 @@
   "version": "0.0.1",
   "description": "Example to show usage with formik",
   "main": "index.js",
-  "scripts": {
-    "start": "npm run devserver",
-    "clean": ".\\node_modules\\.bin\\rimraf -rf public/dist/",
-    "dist": ".\\node_modules\\.bin\\webpack",
-    "build": "webpack --config webpack.config.js",
-    "devserver": ".\\node_modules\\.bin\\webpack-dev-server --content-base public/"
-  },
-  "author": "pstr",
+  "author": "pstrh",
   "license": "MIT",
+  "scripts": {
+    "start": "parcel index.html",
+    "build": "parcel build index.html"
+  },
   "dependencies": {
-    "formik": "1.3.0",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1"
+    "formik": "1.5.8",
+    "react-app-polyfill": "1.0.0"
+  },
+  "alias": {
+    "react": "../../node_modules/react",
+    "react-dom": "../../node_modules/react-dom/profiling",
+    "scheduler/tracing": "../../node_modules/scheduler/tracing-profiling"
   },
   "devDependencies": {
-    "@types/react": "^16.3.16",
-    "@types/react-dom": "^16.0.6",
-    "awesome-typescript-loader": "^5.0.0",
-    "source-map-loader": "^0.2.3",
-    "typescript": "^2.9.1",
-    "webpack": "^4.11.1",
-    "webpack-cli": "^3.0.2",
-    "webpack-dev-server": "^3.1.9",
-    "prettier": "^1.12.0"
-  },
-  "prettier": {
-    "printWidth": 150,
-    "tabWidth": 4,
-    "arrowParens": "always"
+    "@types/react": "16.8.15",
+    "@types/react-dom": "16.8.4",
+    "parcel": "1.12.3",
+    "typescript": "3.4.5"
   }
 }
