--- examples/formik/public/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head lang="de">
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="normalize.css">
-    <title>TypeScript Workspace</title>
-  </head>
-
-  <body style="margin: 50px">
-    <div id="mount"></div>
-  </body>
-
-  <script src="dist/app.js"></script>
-</html>
