+++ examples/formik/tsconfig.json
@@ -1,14 +1,19 @@
 {
   "compilerOptions": {
-    "lib": ["es2015", "es2016.array.include", "dom"],
+    "allowSyntheticDefaultImports": false,
     "target": "es5",
-    "strict": true,
-    "allowJs": true,
+    "module": "commonjs",
     "jsx": "react",
-    "module": "ESNext",
     "moduleResolution": "node",
-    "experimentalDecorators": true
-  },
-  "include": ["./src/**/*"],
-  "exclude": ["node_modules"]
+    "noImplicitAny": false,
+    "noUnusedLocals": false,
+    "noUnusedParameters": false,
+    "removeComments": true,
+    "strictNullChecks": true,
+    "preserveConstEnums": true,
+    "sourceMap": true,
+    "lib": ["es2015", "es2016", "dom"],
+    "baseUrl": ".",
+    "types": ["node"]
+  }
 }
