+++ package.json
@@ -18,39 +18,36 @@
   "author": "pstrh",
   "license": "MIT",
   "main": "dist/index.js",
+  "module": "dist/react-constraint-validation.esm.js",
   "typings": "dist/index.d.ts",
-  "scripts": {
-    "test": "jest",
-    "test-watch": "jest --watch",
-    "build": "tsc"
-  },
-  "devDependencies": {
-    "@types/jest": "24.0.15",
-    "@types/react": "16.8.23",
-    "@types/react-test-renderer": "16.8.3",
-    "jest": "24.8.0",
-    "react": "16.8.6",
-    "react-test-renderer": "16.8.6",
-    "ts-jest": "24.0.2",
-    "typescript": "3.5.3"
-  },
-  "jest": {
-    "transform": {
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "testRegex": "(/__test__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ]
-  },
   "files": [
     "dist"
   ],
+  "scripts": {
+    "start": "tsdx watch",
+    "build": "tsdx build",
+    "test": "tsdx test --env=jsdom",
+    "lint": "tsdx lint"
+  },
+  "prettier": {
+    "printWidth": 80,
+    "semi": true,
+    "tabWidth": 4,
+    "trailingComma": "es5"
+  },
+  "devDependencies": {
+    "@types/jest": "24.0.18",
+    "@types/react": "16.9.3",
+    "@types/react-dom": "16.9.1",
+    "@types/react-test-renderer": "16.9.0",
+    "husky": "3.0.5",
+    "react": "16.9.0",
+    "react-dom": "16.9.0",
+    "react-test-renderer": "16.9.0",
+    "tsdx": "0.9.2",
+    "tslib": "1.10.0",
+    "typescript": "3.6.3"
+  },
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
   }
